requirements.txt.txt
numpy==1.9.2
scikit-learn==0.18
pandas
pillow==9.2.0